requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==5.2
django-filter==25.1
django-model-utils==5.0.0
django-notifications-hq==1.8.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
jsonfield==3.1.0
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
routing==0.2.0
sqlparse==0.5.3
swapper==1.4.0
tzdata==2025.2
pytest~=8.3.5
factory_boy~=3.3.3